+++ packages/stream_chat_flutter/lib/src/scroll_view/draft_scroll_view/stream_draft_list_view.dart
@@ -317,7 +317,7 @@ class StreamDraftListView extends StatelessWidget {
                 padding: const EdgeInsets.all(8),
                 child: StreamScrollViewEmptyWidget(
                   emptyIcon: Icon(
-                    context.streamIcons.edit32,
+                    context.streamIcons.edit,
                     size: 148,
                     color: chatThemeData.colorTheme.disabled,
                   ),