+++ packages/stream_chat_flutter/lib/src/scroll_view/photo_gallery/stream_photo_gallery.dart
@@ -367,7 +367,7 @@ class StreamPhotoGallery extends StatelessWidget {
                 padding: const EdgeInsets.all(8),
                 child: StreamScrollViewEmptyWidget(
                   emptyIcon: Icon(
-                    context.streamIcons.image32,
+                    context.streamIcons.imageLarge,
                     size: 148,
                     color: chatThemeData.colorTheme.disabled,
                   ),