+++ packages/stream_chat_flutter/lib/src/scroll_view/stream_scroll_view_load_more_error.dart
@@ -77,7 +77,7 @@ class StreamScrollViewLoadMoreError extends StatelessWidget {
       child:
           this.errorIcon ??
           Icon(
-            context.streamIcons.retry20,
+            context.streamIcons.retry,
             color: Colors.white,
           ),
     );