+++ packages/stream_chat_flutter/lib/src/scroll_view/thread_scroll_view/stream_unread_threads_banner.dart
@@ -75,7 +75,7 @@ class StreamUnreadThreadsBanner extends StatelessWidget {
               ),
             ),
             Icon(
-              context.streamIcons.refresh20,
+              context.streamIcons.refresh,
               color: theme.colorTheme.barsBg,
             ),
           ],