+++ sample_app/lib/pages/group_info_screen.dart
@@ -10,7 +10,6 @@ import 'package:sample_app/pages/channel_file_display_screen.dart';
 import 'package:sample_app/pages/channel_media_display_screen.dart';
 import 'package:sample_app/pages/pinned_messages_screen.dart';
 import 'package:sample_app/routes/routes.dart';
-import 'package:sample_app/utils/localizations.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
 
 class GroupInfoScreen extends StatefulWidget {
@@ -137,7 +136,7 @@ class _GroupInfoScreenState extends State<GroupInfoScreen> {
                   builder: (context, state) {
                     if (!state.hasData) {
                       return Text(
-                        AppLocalizations.of(context).loading,
+                        'Loading...',
                         style: TextStyle(
                           color: StreamChatTheme.of(context).colorTheme.textHighEmphasis,
                           fontSize: 16,
@@ -167,7 +166,7 @@ class _GroupInfoScreenState extends State<GroupInfoScreen> {
                   height: 3,
                 ),
                 Text(
-                  '${channel.memberCount} ${AppLocalizations.of(context).members}, ${snapshot.data?.where((e) => e.user!.online).length ?? 0} ${AppLocalizations.of(context).online}',
+                  '${channel.memberCount} Members, ${snapshot.data?.where((e) => e.user!.online).length ?? 0} Online',
                   style: TextStyle(
                     color: StreamChatTheme.of(context).colorTheme.textHighEmphasis.withOpacity(0.5),
                     fontSize: 12,
@@ -186,7 +185,7 @@ class _GroupInfoScreenState extends State<GroupInfoScreen> {
                     child: Padding(
                       padding: const EdgeInsets.all(8),
                       child: Icon(
-                        context.streamIcons.userAdd20,
+                        context.streamIcons.userAdd,
                         color: StreamChatTheme.of(context).colorTheme.accentPrimary,
                       ),
                     ),
@@ -282,7 +281,7 @@ class _GroupInfoScreenState extends State<GroupInfoScreen> {
                           Padding(
                             padding: const EdgeInsets.all(8),
                             child: Text(
-                              member.userId == channel.createdBy?.id ? AppLocalizations.of(context).owner : '',
+                              member.userId == channel.createdBy?.id ? 'Owner' : '',
                               style: TextStyle(
                                 color: StreamChatTheme.of(context).colorTheme.textHighEmphasis.withOpacity(0.5),
                               ),
@@ -320,7 +319,7 @@ class _GroupInfoScreenState extends State<GroupInfoScreen> {
                           Padding(
                             padding: const EdgeInsets.symmetric(horizontal: 21, vertical: 12),
                             child: Icon(
-                              context.streamIcons.chevronDown20,
+                              context.streamIcons.chevronDown,
                               color: StreamChatTheme.of(context).colorTheme.textLowEmphasis,
                             ),
                           ),
@@ -330,7 +329,7 @@ class _GroupInfoScreenState extends State<GroupInfoScreen> {
                               mainAxisAlignment: MainAxisAlignment.center,
                               children: [
                                 Text(
-                                  '${members.length - groupMembersLength} ${AppLocalizations.of(context).more}',
+                                  '${members.length - groupMembersLength} more',
                                   style: TextStyle(color: StreamChatTheme.of(context).colorTheme.textLowEmphasis),
                                 ),
                               ],
@@ -365,7 +364,7 @@ class _GroupInfoScreenState extends State<GroupInfoScreen> {
             Padding(
               padding: const EdgeInsets.all(7),
               child: Text(
-                AppLocalizations.of(context).name.toUpperCase(),
+                'Name'.toUpperCase(),
                 style: StreamChatTheme.of(context).textTheme.footnote.copyWith(
                   color: StreamChatTheme.of(context).colorTheme.textHighEmphasis.withOpacity(0.5),
                 ),
@@ -381,7 +380,7 @@ class _GroupInfoScreenState extends State<GroupInfoScreen> {
                 controller: _nameController,
                 cursorColor: StreamChatTheme.of(context).colorTheme.textHighEmphasis,
                 decoration: InputDecoration.collapsed(
-                  hintText: AppLocalizations.of(context).addAGroupName,
+                  hintText: 'Add a group name',
                   hintStyle: StreamChatTheme.of(context).textTheme.bodyBold.copyWith(
                     color: StreamChatTheme.of(context).colorTheme.textHighEmphasis.withOpacity(0.5),
                   ),
@@ -397,7 +396,7 @@ class _GroupInfoScreenState extends State<GroupInfoScreen> {
                 mainAxisSize: MainAxisSize.min,
                 children: [
                   IconButton(
-                    icon: Icon(context.streamIcons.xmark16),
+                    icon: Icon(context.streamIcons.xmark),
                     onPressed: () {
                       setState(() {
                         _nameController.text = _getChannelName(
@@ -412,7 +411,7 @@ class _GroupInfoScreenState extends State<GroupInfoScreen> {
                   ),
                   IconButton(
                     color: StreamChatTheme.of(context).colorTheme.accentPrimary,
-                    icon: Icon(context.streamIcons.checkmark20),
+                    icon: Icon(context.streamIcons.checkmark),
                     onPressed: () {
                       try {
                         channel.update({
@@ -439,32 +438,32 @@ class _GroupInfoScreenState extends State<GroupInfoScreen> {
       children: [
         if (channel.canMuteChannel)
           _GroupInfoToggle(
-            title: AppLocalizations.of(context).muteGroup,
-            icon: context.streamIcons.mute20,
+            title: 'Mute group',
+            icon: context.streamIcons.mute,
             channelStream: channel.isMutedStream,
             localNotifier: mutedBool,
             onTurnOff: channel.unmute,
             onTurnOn: channel.mute,
           ),
         _GroupInfoToggle(
-          title: AppLocalizations.of(context).pinGroup,
-          icon: context.streamIcons.pin20,
+          title: 'Pin group',
+          icon: context.streamIcons.pin,
           channelStream: channel.isPinnedStream,
           localNotifier: isPinned,
           onTurnOff: channel.unpin,
           onTurnOn: channel.pin,
         ),
         _GroupInfoToggle(
-          title: AppLocalizations.of(context).archiveGroup,
-          icon: context.streamIcons.save20,
+          title: 'Archive group',
+          icon: context.streamIcons.save,
           channelStream: channel.isArchivedStream,
           localNotifier: isArchived,
           onTurnOff: channel.unarchive,
           onTurnOn: channel.archive,
         ),
         _GroupInfoListTile(
-          title: AppLocalizations.of(context).pinnedMessages,
-          icon: context.streamIcons.pin20,
+          title: 'Pinned Messages',
+          icon: context.streamIcons.pin,
           iconSize: 24,
           iconPadding: 16,
           onTap: () {
@@ -482,8 +481,8 @@ class _GroupInfoScreenState extends State<GroupInfoScreen> {
           },
         ),
         _GroupInfoListTile(
-          title: AppLocalizations.of(context).photosAndVideos,
-          icon: context.streamIcons.image32,
+          title: 'Photos & Videos',
+          icon: context.streamIcons.imageLarge,
           iconSize: 32,
           iconPadding: 12,
           onTap: () {
@@ -503,8 +502,8 @@ class _GroupInfoScreenState extends State<GroupInfoScreen> {
           },
         ),
         _GroupInfoListTile(
-          title: AppLocalizations.of(context).files,
-          icon: context.streamIcons.file32,
+          title: 'Files',
+          icon: context.streamIcons.file,
           iconSize: 32,
           iconPadding: 12,
           onTap: () {
@@ -527,12 +526,12 @@ class _GroupInfoScreenState extends State<GroupInfoScreen> {
           StreamOptionListTile(
             tileColor: StreamChatTheme.of(context).colorTheme.appBg,
             separatorColor: StreamChatTheme.of(context).colorTheme.disabled,
-            title: AppLocalizations.of(context).leaveGroup,
+            title: 'Leave Group',
             titleTextStyle: StreamChatTheme.of(context).textTheme.body,
             leading: Padding(
               padding: const EdgeInsets.symmetric(horizontal: 16),
               child: Icon(
-                context.streamIcons.userRemove20,
+                context.streamIcons.userRemove,
                 size: 24,
                 color: StreamChatTheme.of(context).colorTheme.textHighEmphasis.withOpacity(0.5),
               ),
@@ -547,12 +546,12 @@ class _GroupInfoScreenState extends State<GroupInfoScreen> {
               final router = GoRouter.of(context);
               final res = await showConfirmationBottomSheet(
                 context,
-                title: AppLocalizations.of(context).leaveConversation,
-                okText: AppLocalizations.of(context).leave.toUpperCase(),
-                question: AppLocalizations.of(context).leaveConversationAreYouSure,
-                cancelText: AppLocalizations.of(context).cancel.toUpperCase(),
+                title: 'Leave conversation',
+                okText: 'Leave'.toUpperCase(),
+                question: 'Are you sure you want to leave this conversation?',
+                cancelText: 'Cancel'.toUpperCase(),
                 icon: Icon(
-                  context.streamIcons.userRemove20,
+                  context.streamIcons.userRemove,
                   color: StreamChatTheme.of(context).colorTheme.accentError,
                 ),
               );
@@ -614,12 +613,12 @@ class _GroupInfoScreenState extends State<GroupInfoScreen> {
                                       Padding(
                                         padding: const EdgeInsets.all(24),
                                         child: Icon(
-                                          context.streamIcons.search32,
+                                          context.streamIcons.search,
                                           size: 96,
                                           color: StreamChatTheme.of(context).colorTheme.textLowEmphasis,
                                         ),
                                       ),
-                                      Text(AppLocalizations.of(context).noUserMatchesTheseKeywords),
+                                      const Text('No user matches these keywords...'),
                                     ],
                                   ),
                                 ),
@@ -654,13 +653,13 @@ class _GroupInfoScreenState extends State<GroupInfoScreen> {
               cursorColor: theme.colorTheme.textHighEmphasis,
               autofocus: true,
               decoration: InputDecoration(
-                hintText: AppLocalizations.of(context).search,
+                hintText: 'Search',
                 hintStyle: theme.textTheme.body.copyWith(
                   color: theme.colorTheme.textLowEmphasis,
                 ),
                 prefixIconConstraints: BoxConstraints.tight(const Size(40, 24)),
                 prefixIcon: Icon(
-                  context.streamIcons.search20,
+                  context.streamIcons.search,
                   color: theme.colorTheme.textHighEmphasis,
                   size: 24,
                 ),
@@ -682,7 +681,7 @@ class _GroupInfoScreenState extends State<GroupInfoScreen> {
           ),
         ),
         IconButton(
-          icon: Icon(context.streamIcons.xmark16),
+          icon: Icon(context.streamIcons.xmark),
           color: theme.colorTheme.textHighEmphasis,
           onPressed: () => Navigator.pop(context),
         ),
@@ -736,11 +735,11 @@ class _GroupInfoScreenState extends State<GroupInfoScreen> {
                     _buildModalListTile(
                       context,
                       Icon(
-                        context.streamIcons.user20,
+                        context.streamIcons.user,
                         color: StreamChatTheme.of(context).colorTheme.textLowEmphasis,
                         size: 24,
                       ),
-                      AppLocalizations.of(context).viewInfo,
+                      'View info',
                       () async {
                         final client = StreamChat.of(context).client;
                         final router = GoRouter.of(context);
@@ -768,11 +767,11 @@ class _GroupInfoScreenState extends State<GroupInfoScreen> {
                     _buildModalListTile(
                       context,
                       Icon(
-                        context.streamIcons.messageBubble20,
+                        context.streamIcons.messageBubble,
                         color: StreamChatTheme.of(context).colorTheme.textLowEmphasis,
                         size: 24,
                       ),
-                      AppLocalizations.of(context).message,
+                      'Message',
                       () async {
                         final client = StreamChat.of(context).client;
                         final router = GoRouter.of(context);
@@ -799,19 +798,19 @@ class _GroupInfoScreenState extends State<GroupInfoScreen> {
                     _buildModalListTile(
                       context,
                       Icon(
-                        context.streamIcons.userRemove20,
+                        context.streamIcons.userRemove,
                         color: StreamChatTheme.of(context).colorTheme.accentError,
                         size: 24,
                       ),
-                      AppLocalizations.of(context).removeFromGroup,
+                      'Remove From Group',
                       () async {
                         final router = GoRouter.of(context);
                         final res = await showConfirmationBottomSheet(
                           context,
-                          title: AppLocalizations.of(context).removeMember,
-                          okText: AppLocalizations.of(context).remove.toUpperCase(),
-                          question: AppLocalizations.of(context).removeMemberAreYouSure,
-                          cancelText: AppLocalizations.of(context).cancel.toUpperCase(),
+                          title: 'Remove member',
+                          okText: 'Remove'.toUpperCase(),
+                          question: 'Are you sure you want to remove this member?',
+                          cancelText: 'Cancel'.toUpperCase(),
                         );
 
                         if (res == true) {
@@ -824,11 +823,11 @@ class _GroupInfoScreenState extends State<GroupInfoScreen> {
                   _buildModalListTile(
                     context,
                     Icon(
-                      context.streamIcons.xmark16,
+                      context.streamIcons.xmark,
                       color: StreamChatTheme.of(context).colorTheme.textLowEmphasis,
                       size: 24,
                     ),
-                    AppLocalizations.of(context).cancel,
+                    'Cancel',
                     () {
                       Navigator.pop(context);
                     },
@@ -856,12 +855,12 @@ class _GroupInfoScreenState extends State<GroupInfoScreen> {
     if (otherMember != null) {
       if (otherMember.online) {
         alternativeWidget = Text(
-          AppLocalizations.of(context).online,
+          'Online',
           style: TextStyle(color: StreamChatTheme.of(context).colorTheme.textHighEmphasis.withOpacity(0.5)),
         );
       } else {
         alternativeWidget = Text(
-          '${AppLocalizations.of(context).lastSeen} ${Jiffy.parseFromDateTime(otherMember.lastActive!).fromNow()}',
+          'Last seen ${Jiffy.parseFromDateTime(otherMember.lastActive!).fromNow()}',
           style: TextStyle(color: StreamChatTheme.of(context).colorTheme.textHighEmphasis.withOpacity(0.5)),
         );
       }
@@ -933,7 +932,7 @@ class _GroupInfoScreenState extends State<GroupInfoScreen> {
         title =
             '${currentMembers.map((e) => e.user!.name).join(', ')} ${exceedingMembers > 0 ? '+ $exceedingMembers' : ''}';
       } else {
-        title = AppLocalizations.of(context).noTitle;
+        title = 'No title';
       }
     } else {
       title = extraData['name'];
@@ -943,13 +942,13 @@ class _GroupInfoScreenState extends State<GroupInfoScreen> {
 
   String _getLastSeen(User user) {
     if (user.online) {
-      return AppLocalizations.of(context).online;
+      return 'Online';
     } else {
       if (user.lastActive == null) {
         return '';
       }
 
-      return '${AppLocalizations.of(context).lastSeen} ${Jiffy.parseFromDateTime(user.lastActive!).fromNow()}';
+      return 'Last seen ${Jiffy.parseFromDateTime(user.lastActive!).fromNow()}';
     }
   }
 }
@@ -1046,7 +1045,7 @@ class _GroupInfoListTile extends StatelessWidget {
         ),
       ),
       trailing: Icon(
-        context.streamIcons.chevronRight20,
+        context.streamIcons.chevronRight,
         color: StreamChatTheme.of(context).colorTheme.textLowEmphasis,
       ),
       onTap: onTap,